+++ test/highlight/javascript/regular.js
@@ -1,5 +1,5 @@
 // Template strings
-let who = 'world';
+const who = 'world';
 console.log(`Hello, ${who}`);
 
 // Nested object
@@ -10,3 +10,10 @@ let some_object = {
 		}
 	}
 };
+
+// Subscript expressions
+const zeroes = [0];
+console.log(zeroes[zeroes[zeroes[0]]])
+
+// Parenthesized expressions
+console.log(1 + (2 + (3 + 4)))