+++ test/highlight/typescript/regular.ts
@@ -32,7 +32,7 @@ class Person {
 }
 
 // Template strings
-let who = 'world';
+const who = 'world';
 console.log(`Hello, ${who}`);
 
 // Nested object
@@ -43,3 +43,10 @@ let some_object = {
 		}
 	}
 };
+
+// Subscript expressions
+const zeroes = [0];
+console.log(zeroes[zeroes[zeroes[0]]])
+
+// Parenthesized expressions
+console.log(1 + (2 + (3 + (4 + (5 + 6)))))